restructure DLT
