+++ nhlPredict/src/ingestion/02-silver-transform.py
@@ -16,13 +16,10 @@ import sys
 
 sys.path.append(spark.conf.get("bundle.sourcePath", "."))
 
-import glob
 from datetime import date
 from pyspark.sql.functions import *
 from pyspark.sql.window import Window
-from pyspark.sql import DataFrame
 from utils.ingestionHelper import (
-    download_unzip_and_save_as_table,
     select_rename_columns,
     select_rename_game_columns,
     get_day_of_week,
@@ -46,245 +43,6 @@ season_list = [2023, 2024]
 
 # Get current date
 today_date = date.today()
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC #### Ingesting of Raw Data - Bronze
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_shots_2023_v2
-# @dlt.table(name="bronze_shots_2023_v2", comment="Raw Ingested NHL data on Shots in 2023")
-# def ingest_shot_data():
-#     shots_file_path = download_unzip_and_save_as_table(
-#         shots_url, tmp_base_path, "shots_2023", file_format=".zip"
-#     )
-#     return spark.read.format("csv").option("header", "true").load(shots_file_path)
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_teams_2023_v2
-# @dlt.table(name="bronze_teams_2023_v2", comment="Raw Ingested NHL data on Teams in 2023")
-# def ingest_teams_data():
-#     teams_file_path = download_unzip_and_save_as_table(
-#         teams_url, tmp_base_path, "teams_2023", file_format=".csv"
-#     )
-#     return (
-#         spark.read.format("csv")
-#         .option("header", "true")
-#         .option("inferSchema", "true")
-#         .load(teams_file_path)
-#     )
-#     # return spark.table("lr_nhl_demo.dev.bronze_teams_2023")
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_skaters_2023_v2
-
-
-@dlt.table(
-    name="bronze_skaters_2023_v2", comment="Raw Ingested NHL data on skaters in 2023"
-)
-def ingest_skaters_data():
-    skaters_file_path = download_unzip_and_save_as_table(
-        skaters_url, tmp_base_path, "skaters_2023", file_format=".csv"
-    )
-
-    skaters_df = (
-        spark.read.format("csv")
-        .option("header", "true")
-        .option("inferSchema", "true")
-        .load(skaters_file_path)
-    )
-    # Add 'game_' before each column name except for 'team' and 'player'
-    skater_columns = skaters_df.columns
-    for column in skater_columns:
-        if column not in [
-            "situation",
-            "season",
-            "team",
-            "name",
-            "playerId",
-        ]:
-            skaters_df = skaters_df.withColumnRenamed(column, f"player_{column}")
-
-    return skaters_df
-
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_lines_2023_v2
-# @dlt.table(name="bronze_lines_2023_v2", comment="Raw Ingested NHL data on lines in 2023")
-# def ingest_lines_data():
-#     lines_file_path = download_unzip_and_save_as_table(
-#         lines_url, tmp_base_path, "lines_2023", file_format=".csv"
-#     )
-#     # return (
-#     #   spark.readStream.format("cloudFiles")
-#     #     .option("cloudFiles.format", "csv")
-#     #     .option("cloudFiles.inferColumnTypes", "true")
-#     #     .option("header", "true")
-#     #     .load(f"{tmp_base_path}lines_2023/")
-#     #   )
-#     return (
-#         spark.read.format("csv")
-#         .option("header", "true")
-#         .option("inferSchema", "true")
-#         .load(lines_file_path)
-#     )
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_games_historical_v2
-
-
-@dlt.table(
-    name="bronze_games_historical_v2",
-    comment="Raw Ingested NHL data on games from 2008 - Present",
-    table_properties={"quality": "bronze"},
-)
-def ingest_games_data():
-    games_file_path = download_unzip_and_save_as_table(
-        games_url, tmp_base_path, "games_historical", file_format=".csv"
-    )
-    return (
-        spark.read.format("csv")
-        .option("header", "true")
-        .option("inferSchema", "true")
-        .load(games_file_path)
-    )
-
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_schedule_2023_v2
-
-
-@dlt.table(
-    name="bronze_schedule_2023_v2",
-    table_properties={"quality": "bronze"},
-)
-def ingest_schedule_data():
-    # TO DO : make live https://media.nhl.com/site/vasset/public/attachments/2023/06/17233/2023-24%20Official%20NHL%20Schedule%20(by%20Day).xlsx
-    return spark.table("lr_nhl_demo.dev.2024_25_official_nhl_schedule_by_day")
-
-
-# @dlt.table(
-#     name="bronze_schedule_2023_v2",
-#     table_properties={"quality": "bronze"},
-# )
-# def ingest_schedule_data():
-#     # TO DO : make live https://media.nhl.com/site/vasset/public/attachments/2023/06/17233/2023-24%20Official%20NHL%20Schedule%20(by%20Day).xlsx
-#     return spark.table("lr_nhl_demo.dev.2023_24_official_nhl_schedule_by_day")
-
-# COMMAND ----------
-
-# DBTITLE 1,bronze_player_game_stats_v2
-
-
-# @dlt.expect_or_drop("team is not null", "team IS NOT NULL")
-# @dlt.expect_or_drop("season is not null", "season IS NOT NULL")
-# @dlt.expect_or_drop("situation is not null", "situation IS NOT NULL")
-# @dlt.expect_or_drop("playerID is not null", "playerID IS NOT NULL")
-@dlt.table(
-    name="bronze_player_game_stats_v2",
-    comment="Game by Game Stats for each player in the skaters table",
-    table_properties={"quality": "bronze"},
-)
-def ingest_games_data():
-    if one_time_load == "true":
-        skaters_2023_id = (
-            dlt.read("bronze_skaters_2023_v2").select("playerId").distinct()
-        )
-        print("Ingesting player game by game stats")
-
-        # Get Playoff teams and players
-        playoff_teams_list = (
-            dlt.read("bronze_games_historical_v2")
-            .select("team")
-            .filter((col("playoffGame") == 1) & (col("season").isin([2023, 2024])))
-            .distinct()
-            .collect()
-        )
-        playoff_teams = [row.team for row in playoff_teams_list]
-        playoff_skaters_2023_id = (
-            dlt.read("bronze_skaters_2023_v2")
-            .select("playerId")
-            .filter(col("team").isin(playoff_teams))
-            .distinct()
-        )
-
-        for row in skaters_2023_id.collect():
-            playerId = str(row["playerId"])
-            games_file_path = download_unzip_and_save_as_table(
-                player_games_url + playerId + ".csv",
-                tmp_base_path,
-                playerId,
-                file_format=".csv",
-                game_by_game=True,
-            )
-
-        # Check if player is in Playoffs, if so bring playoff stats
-        if len(playoff_teams) > 0:
-            for row in playoff_skaters_2023_id.collect():
-                playoff_playerId = str(row["playerId"])
-                playoff_games_file_path = download_unzip_and_save_as_table(
-                    player_playoff_games_url + playoff_playerId + ".csv",
-                    tmp_base_path,
-                    playoff_playerId,
-                    file_format=".csv",
-                    game_by_game_playoffs=True,
-                )
-
-    regular_season_stats_path = "/Volumes/lr_nhl_demo/dev/player_game_stats/*.csv"
-    playoff_season_stats_path = (
-        "/Volumes/lr_nhl_demo/dev/player_game_stats_playoffs/*.csv"
-    )
-
-    # Check for CSV files
-    reg_csv_files = glob.glob(regular_season_stats_path)
-    playoff_csv_files = glob.glob(playoff_season_stats_path)
-
-    if reg_csv_files:
-        regular_season_stats = (
-            spark.read.format("csv")
-            .option("header", "true")
-            .option("inferSchema", "true")
-            .load(regular_season_stats_path)
-        ).filter(col("season").isin(season_list))
-    else:
-        print("No CSV files found for Regular Season. Skipping...")
-        regular_season_stats = (
-            spark.read.format("csv")
-            .options(header="true")
-            .load(
-                "/Volumes/lr_nhl_demo/dev/player_game_stats/8477493.csv"
-            )  # fix this to not be static file
-        )
-
-    if playoff_csv_files:
-        playoff_season_stats = (
-            spark.read.format("csv")
-            .option("header", "true")
-            .option("inferSchema", "true")
-            .load(playoff_season_stats_path)
-        ).filter(col("season").isin(season_list))
-    else:
-        print("No CSV files found for Playoffs. Skipping...")
-        playoff_season_stats = (
-            spark.read.format("csv")
-            .options(header="true")
-            .load("/Volumes/lr_nhl_demo/dev/player_game_stats_playoffs/8477493.csv")
-        )
-
-    return regular_season_stats.union(playoff_season_stats)
-
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC
 
 # COMMAND ----------
 
@@ -1429,657 +1187,3 @@ def clean_rank_players():
 #     )
 
 #     return shots_filtered
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC #### Aggregations - Gold
-
-# COMMAND ----------
-
-# DBTITLE 1,gold_player_stats_v2
-
-
-# @dlt.expect("playerId is not null", "playerId IS NOT NULL")
-# @dlt.expect("shooterName is not null", "shooterName IS NOT NULL")
-@dlt.table(
-    name="gold_player_stats_v2",
-    # comment="Raw Ingested NHL data on games from 2008 - Present",
-    table_properties={"quality": "gold"},
-)
-def aggregate_games_data():
-    gold_shots_date = (
-        dlt.read("silver_games_schedule_v2")
-        .select(
-            "team",
-            "gameId",
-            "season",
-            "home_or_away",
-            "gameDate",
-            "playerTeam",
-            "opposingTeam",
-        )
-        .join(
-            dlt.read("silver_players_ranked").drop(
-                "teamGamesPlayedRolling",
-                "teamMatchupPlayedRolling",
-                "isPlayoffGame",
-                "game_Total_penaltiesAgainst",
-                "rolling_game_Total_penaltiesAgainst",
-                "rolling_per_game_game_Total_penaltiesAgainst",
-                "rank_rolling_game_Total_penaltiesAgainst",
-                "perc_rank_rolling_game_Total_penaltiesAgainst",
-            ),
-            how="left",
-            on=[
-                "playerTeam",
-                "gameId",
-                "gameDate",
-                "opposingTeam",
-                "season",
-                "home_or_away",
-            ],
-        )
-    )
-
-    if str(today_date) >= str(
-        dlt.read("bronze_schedule_2023_v2").select(min("DATE")).first()[0]
-    ):  # SETTING TO ALWAYS BE TRUE FOR NOW
-        player_index_2023 = (
-            dlt.read("bronze_skaters_2023_v2")
-            .select("playerId", "season", "team", "name")
-            .filter(col("situation") == "all")
-            .distinct()
-            .unionByName(
-                dlt.read("bronze_skaters_2023_v2")
-                .select("playerId", "season", "team", "name")
-                .filter(col("situation") == "all")
-                .withColumn("season", lit(2024))
-                .distinct()
-            )
-        )
-    else:
-        player_index_2023 = (
-            dlt.read("bronze_skaters_2023_v2")
-            .select("playerId", "season", "team", "name")
-            .filter(col("situation") == "all")
-            .distinct()
-        )
-
-    player_game_index_2023 = (
-        dlt.read("silver_games_schedule_v2")
-        .select(
-            "team",
-            "gameId",
-            "season",
-            "home_or_away",
-            "gameDate",
-            "playerTeam",
-            "opposingTeam",
-        )
-        .join(player_index_2023, how="left", on=["team", "season"])
-        .select("team", "playerId", "season", "name")
-        .distinct()
-        .withColumnRenamed("name", "shooterName")
-    ).alias("player_game_index_2023")
-
-    silver_games_schedule = (
-        dlt.read("silver_games_schedule_v2")
-        .select(
-            "team",
-            "gameId",
-            "season",
-            "home_or_away",
-            "gameDate",
-            "playerTeam",
-            "opposingTeam",
-        )
-        .alias("silver_games_schedule")
-    )
-
-    for col_name in player_game_index_2023.columns:
-        player_game_index_2023 = player_game_index_2023.withColumnRenamed(
-            col_name, "index_" + col_name
-        )
-
-    player_game_index_2023 = player_game_index_2023.alias("player_game_index_2023")
-
-    upcoming_games_player_index = silver_games_schedule.filter(
-        col("gameId").isNull()
-    ).join(
-        player_game_index_2023,
-        how="left",
-        on=[col("index_team") == col("team"), col("index_season") == col("season")],
-    )
-
-    gold_shots_date_final = (
-        gold_shots_date.join(
-            upcoming_games_player_index.drop("gameId"),
-            how="left",
-            on=[
-                "team",
-                "season",
-                "home_or_away",
-                "gameDate",
-                "playerTeam",
-                "opposingTeam",
-            ],
-        )
-        .withColumn(
-            "playerId",
-            when(col("playerId").isNull(), col("index_playerId")).otherwise(
-                col("playerId")
-            ),
-        )
-        .withColumn(
-            "shooterName",
-            when(col("shooterName").isNull(), col("index_shooterName")).otherwise(
-                col("shooterName")
-            ),
-        )
-        .drop("index_season", "index_team", "index_shooterName", "index_playerId")
-    )
-
-    # Define Windows (player last games, and players last matchups)
-    windowSpec = Window.partitionBy("playerId", "playerTeam", "shooterName").orderBy(
-        col("gameDate")
-    )
-    last3WindowSpec = windowSpec.rowsBetween(-2, 0)
-    last7WindowSpec = windowSpec.rowsBetween(-6, 0)
-    matchupWindowSpec = Window.partitionBy(
-        "playerId", "playerTeam", "shooterName", "opposingTeam"
-    ).orderBy(col("gameDate"))
-    matchupLast3WindowSpec = matchupWindowSpec.rowsBetween(-2, 0)
-    matchupLast7WindowSpec = matchupWindowSpec.rowsBetween(-6, 0)
-
-    reorder_list = [
-        "gameDate",
-        "gameId",
-        "season",
-        "position",
-        "home_or_away",
-        "isHome",
-        "isPlayoffGame",
-        "playerTeam",
-        "opposingTeam",
-        "playerId",
-        "shooterName",
-        "DAY",
-        "DATE",
-        "dummyDay",
-        "AWAY",
-        "HOME",
-        "team",
-        "homeTeamCode",
-        "awayTeamCode",
-        "playerGamesPlayedRolling",
-        "playerMatchupPlayedRolling",
-    ]
-
-    # Create a window specification
-    gameCountWindowSpec = (
-        Window.partitionBy("playerId", "playerTeam", "season")
-        .orderBy("gameDate")
-        .rowsBetween(Window.unboundedPreceding, 0)
-    )
-    matchupCountWindowSpec = (
-        Window.partitionBy("playerId", "playerTeam", "opposingTeam", "season")
-        .orderBy("gameDate")
-        .rowsBetween(Window.unboundedPreceding, 0)
-    )
-
-    # Apply the count function within the window
-    gold_shots_date_count = gold_shots_date_final = (
-        gold_shots_date_final.withColumn(
-            "playerGamesPlayedRolling", count("gameId").over(gameCountWindowSpec)
-        )
-        .withColumn(
-            "playerMatchupPlayedRolling", count("gameId").over(matchupCountWindowSpec)
-        )
-        .withColumn(
-            "player_SOG%_PP",
-            round(
-                when(
-                    col("player_Total_shotsOnGoal") != 0,
-                    col("player_PP_shotsOnGoal") / col("player_Total_shotsOnGoal"),
-                ).otherwise(None),
-                2,
-            ),
-        )
-        .withColumn(
-            "player_SOG%_EV",
-            round(
-                when(
-                    col("player_Total_shotsOnGoal") != 0,
-                    col("player_EV_shotsOnGoal") / col("player_Total_shotsOnGoal"),
-                ).otherwise(None),
-                2,
-            ),
-        )
-    )
-
-    columns_to_iterate = [
-        col for col in gold_shots_date_count.columns if col not in reorder_list
-    ]
-
-    # Create a list of column expressions for lag and averages
-    column_exprs = [
-        col(c) for c in gold_shots_date_count.columns
-    ]  # Start with all existing columns
-
-    # might be nulls still for players that have just switched teams, will get imputed in Feat Engineering
-    player_avg_exprs = {
-        col_name: round(lag(col(col_name)).over(windowSpec), 2)
-        for col_name in columns_to_iterate
-    }
-    player_avg3_exprs = {
-        col_name: round(
-            mean(col(col_name)).over(windowSpec.rowsBetween(-3, 1)),
-            2,
-        )
-        for col_name in columns_to_iterate
-    }
-    player_avg7_exprs = {
-        col_name: round(
-            mean(col(col_name)).over(windowSpec.rowsBetween(-7, 1)),
-            2,
-        )
-        for col_name in columns_to_iterate
-    }
-    playerMatch_avg_exprs = {
-        col_name: round(lag(col(col_name)).over(matchupWindowSpec), 2)
-        for col_name in columns_to_iterate
-    }
-    playerMatch_avg3_exprs = {
-        col_name: round(
-            mean(col(col_name)).over(matchupWindowSpec.rowsBetween(-3, 1)),
-            2,
-        )
-        for col_name in columns_to_iterate
-    }
-    playerMatch_avg7_exprs = {
-        col_name: round(
-            mean(col(col_name)).over(matchupWindowSpec.rowsBetween(-7, 1)),
-            2,
-        )
-        for col_name in columns_to_iterate
-    }
-
-    for column_name in columns_to_iterate:
-        player_avg = player_avg_exprs[column_name]
-        player_avg3 = player_avg3_exprs[column_name]
-        player_avg7 = player_avg7_exprs[column_name]
-        matchup_avg = playerMatch_avg_exprs[column_name]
-        matchup_avg3 = playerMatch_avg3_exprs[column_name]
-        matchup_avg7 = playerMatch_avg7_exprs[column_name]
-
-        column_exprs += [
-            when(
-                col("gameId").isNotNull(),
-                round(lag(col(column_name)).over(windowSpec), 2),
-            )
-            .otherwise(player_avg)
-            .alias(f"previous_{column_name}"),
-            when(
-                col("gameId").isNotNull(),
-                round(mean(col(column_name)).over(last3WindowSpec), 2),
-            )
-            .otherwise(round(player_avg3, 2))
-            .alias(f"average_{column_name}_last_3_games"),
-            when(
-                col("gameId").isNotNull(),
-                round(mean(col(column_name)).over(last7WindowSpec), 2),
-            )
-            .otherwise(player_avg7)
-            .alias(f"average_{column_name}_last_7_games"),
-            when(
-                col("gameId").isNotNull(),
-                round(lag(col(column_name)).over(matchupWindowSpec), 2),
-            )
-            .otherwise(matchup_avg)
-            .alias(f"matchup_previous_{column_name}"),
-            when(
-                col("gameId").isNotNull(),
-                round(mean(col(column_name)).over(matchupLast3WindowSpec), 2),
-            )
-            .otherwise(matchup_avg3)
-            .alias(f"matchup_average_{column_name}_last_3_games"),
-            when(
-                col("gameId").isNotNull(),
-                round(mean(col(column_name)).over(matchupLast7WindowSpec), 2),
-            )
-            .otherwise(matchup_avg7)
-            .alias(f"matchup_average_{column_name}_last_7_games"),
-        ]
-
-    # Apply all column expressions at once using select
-    gold_player_stats = gold_shots_date_count.select(*column_exprs)
-
-    assert (
-        gold_player_stats.filter(
-            (col("playerId").isNull()) & (col("playerTeam") != "UTA")
-        ).count()
-        == 0
-    ), f"PlayerId Null Rows {gold_player_stats.filter(col('playerId').isNull()).count()}"
-
-    return gold_player_stats
-
-
-# COMMAND ----------
-
-# DBTITLE 1,gold_game_stats_v2
-
-
-@dlt.table(
-    name="gold_game_stats_v2",
-    # comment="Raw Ingested NHL data on games from 2008 - Present",
-    table_properties={"quality": "gold"},
-)
-def window_gold_game_data():
-    # Define Windows (team last games, and team last matchups)
-    # for each distinct opposingTeam, get the last game {metric}. Metrics are: game_Total_shotsOnGoalAgainst, game_Total_goalsAgainst, game_Total_shotAttemptsAgainst, game_Total_penaltiesAgainst, game_PK_shotsOnGoalAgainst, game_PK_goalsAgainst, game_PK_shotAttemptsAgainst, game_PP_shotsOnGoalFor, game_PP_goalsFor, game_PP_shotAttemptsFor
-
-    windowSpec = Window.partitionBy("playerTeam").orderBy(col("gameDate"))
-    last3WindowSpec = windowSpec.rowsBetween(-2, 0)
-    last7WindowSpec = windowSpec.rowsBetween(-6, 0)
-    opponentWindowSpec = Window.partitionBy("opposingTeam").orderBy(col("gameDate"))
-    opponentLast3WindowSpec = opponentWindowSpec.rowsBetween(-2, 0)
-    opponentLast7WindowSpec = opponentWindowSpec.rowsBetween(-6, 0)
-    matchupWindowSpec = Window.partitionBy("playerTeam", "opposingTeam").orderBy(
-        col("gameDate")
-    )
-    matchupLast3WindowSpec = matchupWindowSpec.rowsBetween(-2, 0)
-    matchupLast7WindowSpec = matchupWindowSpec.rowsBetween(-6, 0)
-
-    reorder_list = [
-        "gameDate",
-        "gameId",
-        "season",
-        "home_or_away",
-        "isHome",
-        "isPlayoffGame",
-        "playerTeam",
-        "opposingTeam",
-        "playerId",
-        "shooterName",
-        "DAY",
-        "DATE",
-        "dummyDay",
-        "AWAY",
-        "HOME",
-        "team",
-        "teamGamesPlayedRolling",
-        "teamMatchupPlayedRolling",
-    ]
-
-    # Apply the count function within the window
-    gold_games_count = dlt.read("silver_games_rankings").drop(
-        "EASTERN", "LOCAL", "homeTeamCode", "awayTeamCode"
-    )
-
-    columns_to_iterate = [
-        col for col in gold_games_count.columns if col not in reorder_list
-    ]
-
-    # Create a list of column expressions for lag and averages
-    column_exprs = [
-        col(c) for c in gold_games_count.columns
-    ]  # Start with all existing columns
-    game_avg_exprs = {
-        col_name: round(median(col(col_name)).over(Window.partitionBy("playerTeam")), 2)
-        for col_name in columns_to_iterate
-    }
-    opponent_game_avg_exprs = {
-        col_name: round(
-            median(col(col_name)).over(Window.partitionBy("opposingTeam")), 2
-        )
-        for col_name in columns_to_iterate
-    }
-    matchup_avg_exprs = {
-        col_name: round(
-            median(col(col_name)).over(
-                Window.partitionBy("playerTeam", "opposingTeam")
-            ),
-            2,
-        )
-        for col_name in columns_to_iterate
-    }
-
-    for column_name in columns_to_iterate:
-        game_avg = game_avg_exprs[column_name]
-        opponent_game_avg = opponent_game_avg_exprs[column_name]
-        matchup_avg = matchup_avg_exprs[column_name]
-        column_exprs += [
-            when(
-                col("teamGamesPlayedRolling") > 1,
-                round(lag(col(column_name)).over(windowSpec), 2),
-            )
-            .otherwise(game_avg)
-            .alias(f"previous_{column_name}"),
-            when(
-                col("teamGamesPlayedRolling") > 3,
-                round(avg(col(column_name)).over(last3WindowSpec), 2),
-            )
-            .otherwise(game_avg)
-            .alias(f"average_{column_name}_last_3_games"),
-            when(
-                col("teamGamesPlayedRolling") > 7,
-                round(avg(col(column_name)).over(last7WindowSpec), 2),
-            )
-            .otherwise(game_avg)
-            .alias(f"average_{column_name}_last_7_games"),
-            when(
-                col("teamGamesPlayedRolling") > 1,
-                round(lag(col(column_name)).over(opponentWindowSpec), 2),
-            )
-            .otherwise(opponent_game_avg)
-            .alias(f"opponent_previous_{column_name}"),
-            when(
-                col("teamGamesPlayedRolling") > 3,
-                round(avg(col(column_name)).over(opponentLast3WindowSpec), 2),
-            )
-            .otherwise(opponent_game_avg)
-            .alias(f"opponent_average_{column_name}_last_3_games"),
-            when(
-                col("teamGamesPlayedRolling") > 7,
-                round(avg(col(column_name)).over(opponentLast7WindowSpec), 2),
-            )
-            .otherwise(opponent_game_avg)
-            .alias(f"opponent_average_{column_name}_last_7_games"),
-            when(
-                col("teamMatchupPlayedRolling") > 1,
-                round(lag(col(column_name)).over(matchupWindowSpec), 2),
-            )
-            .otherwise(matchup_avg)
-            .alias(f"matchup_previous_{column_name}"),
-            when(
-                col("teamMatchupPlayedRolling") > 3,
-                round(avg(col(column_name)).over(matchupLast3WindowSpec), 2),
-            )
-            .otherwise(matchup_avg)
-            .alias(f"matchup_average_{column_name}_last_3_games"),
-            when(
-                col("teamMatchupPlayedRolling") > 7,
-                round(avg(col(column_name)).over(matchupLast7WindowSpec), 2),
-            )
-            .otherwise(matchup_avg)
-            .alias(f"matchup_average_{column_name}_last_7_games"),
-        ]
-
-    # Apply all column expressions at once using select
-    gold_game_stats = gold_games_count.select(*column_exprs).withColumn(
-        "previous_opposingTeam", lag(col("opposingTeam")).over(windowSpec)
-    )
-
-    return gold_game_stats
-
-
-# COMMAND ----------
-
-# DBTITLE 1,gold_merged_stats_v2
-
-
-@dlt.table(
-    name="gold_merged_stats_v2",
-    # comment="Raw Ingested NHL data on games from 2008 - Present",
-    table_properties={"quality": "gold"},
-)
-def merge_player_game_stats():
-    gold_player_stats = dlt.read("gold_player_stats_v2").alias("gold_player_stats")
-    gold_game_stats = dlt.read("gold_game_stats_v2").alias("gold_game_stats")
-    # schedule_2023 = dlt.read("bronze_schedule_2023").alias("schedule_2023")
-
-    gold_merged_stats = gold_game_stats.join(
-        gold_player_stats.drop("gameId"),
-        how="left",
-        on=[
-            "team",
-            "season",
-            "home_or_away",
-            "gameDate",
-            "playerTeam",
-            "opposingTeam",
-        ],
-    ).alias("gold_merged_stats")
-
-    schedule_shots = (
-        gold_merged_stats.drop("EASTERN", "LOCAL", "homeTeamCode", "awayTeamCode")
-        .withColumn("isHome", when(col("home_or_away") == "HOME", 1).otherwise(0))
-        .withColumn(
-            "dummyDay",
-            when(col("DAY") == "Mon", 1)
-            .when(col("DAY") == "Tue", 2)
-            .when(col("DAY") == "Wed", 3)
-            .when(col("DAY") == "Thu", 4)
-            .when(col("DAY") == "Fri", 5)
-            .when(col("DAY") == "Sat", 6)
-            .when(col("DAY") == "Sun", 7)
-            .otherwise(0),
-        )
-        .withColumn("gameId", col("gameId").cast("string"))
-        .withColumn("playerId", col("playerId").cast("string"))
-        .withColumn("gameId", regexp_replace("gameId", "\\.0$", ""))
-        .withColumn("playerId", regexp_replace("playerId", "\\.0$", ""))
-    )
-
-    reorder_list = [
-        "gameDate",
-        "gameId",
-        "season",
-        "position",
-        "home_or_away",
-        "isHome",
-        "isPlayoffGame",
-        "playerTeam",
-        "opposingTeam",
-        "playerId",
-        "shooterName",
-        "DAY",
-        "DATE",
-        "dummyDay",
-        "AWAY",
-        "HOME",
-        "previous_opposingTeam",
-        "playerGamesPlayedRolling",
-        "playerMatchupPlayedRolling",
-        "teamGamesPlayedRolling",
-        "teamMatchupPlayedRolling",
-        "player_Total_shotsOnGoal",
-    ]
-
-    schedule_shots_reordered = schedule_shots.select(
-        *reorder_list,
-        *[col for col in schedule_shots.columns if col not in reorder_list],
-    )
-
-    return schedule_shots_reordered
-
-
-# COMMAND ----------
-
-# DBTITLE 1,gold_model_stats_v2
-
-
-@dlt.table(
-    name="gold_model_stats_v2",
-    # comment="Raw Ingested NHL data on games from 2008 - Present",
-    table_properties={"quality": "gold"},
-)
-def make_model_ready():
-    gold_model_data = dlt.read("gold_merged_stats_v2")
-
-    gameCountWindowSpec = (
-        Window.partitionBy("playerId", "season")
-        .orderBy("gameDate")
-        .rowsBetween(Window.unboundedPreceding, 0)
-    )
-    matchupCountWindowSpec = (
-        Window.partitionBy("playerId", "playerTeam", "opposingTeam", "season")
-        .orderBy("gameDate")
-        .rowsBetween(Window.unboundedPreceding, 0)
-    )
-
-    reorder_list = [
-        "gameDate",
-        "gameId",
-        "season",
-        "position",
-        "home_or_away",
-        "isHome",
-        "isPlayoffGame",
-        "playerTeam",
-        "opposingTeam",
-        "playerId",
-        "shooterName",
-        "DAY",
-        "DATE",
-        "dummyDay",
-        "AWAY",
-        "HOME",
-        "previous_opposingTeam",
-        "playerGamesPlayedRolling",
-        "playerMatchupPlayedRolling",
-        "teamGamesPlayedRolling",
-        "teamMatchupPlayedRolling",
-        "player_Total_shotsOnGoal",
-    ]
-
-    # Create a list of column expressions for lag and averages
-    keep_column_exprs = []  # Start with an empty list
-
-    for column_name in gold_model_data.columns:
-        if (
-            column_name in reorder_list
-            or column_name.startswith("previous")
-            or column_name.startswith("average")
-            or column_name.startswith("matchup")
-            or column_name.startswith("opponent")
-        ):
-            keep_column_exprs.append(col(column_name))
-
-    # Window Spec for calculating sum of 'player_totalTimeOnIceInGame' partitioned by playerId and ordered by gameDate
-    timeOnIceWindowSpec = (
-        Window.partitionBy("playerId")
-        .orderBy("gameDate")
-        .rowsBetween(Window.unboundedPreceding, 0)
-    )
-
-    # Apply all column expressions at once using select
-    gold_model_data = (
-        gold_model_data.select(
-            *keep_column_exprs,
-            round(sum(col("player_Total_icetime")).over(timeOnIceWindowSpec), 2).alias(
-                "rolling_playerTotalTimeOnIceInGame"
-            ),
-        )
-        .withColumn("teamGamesPlayedRolling", count("gameId").over(gameCountWindowSpec))
-        .withColumn(
-            "teamMatchupPlayedRolling", count("gameId").over(matchupCountWindowSpec)
-        )
-        .withColumn(
-            "isPlayoffGame",
-            when(col("teamGamesPlayedRolling") > 82, lit(1)).otherwise(lit(0)),
-        )
-    )
-
-    return gold_model_data
